Move publications into dedicated router
