+++ controllers/publications/views/publication-search.njk
@@ -1,5 +1,4 @@
 {% extends "layouts/main.njk" %}
-
 {% from "components/hero.njk" import hero with context %}
 {% from "components/content-box/macro.njk" import contentBox %}
 {% from "components/breadcrumb-trail/macro.njk" import breadcrumbTrail %}
